alert

+++ app/Http/Controllers/DisciplinaController.php
@@ -5,7 +5,6 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Disciplina;
 use Illuminate\Support\Facades\DB;
-use Image;
 use Storage;
 
 class DisciplinaController extends Controller
@@ -58,28 +57,18 @@ class DisciplinaController extends Controller
 
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function edit($id)
     {
-        $datos = DB::table('disciplinas')->get();
+        $datos = DB::table('disciplinas')
+        ->where('id_disc',$id)
+        ->get();
+
         foreach ($datos as $dato) {
             $disciplina = $dato;
         }
         return view('disciplina.editar_disc')->with('disciplina',$disciplina);
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function update(Request $request, $id)
     {
         if($request->hasFile('foto_disc'))
@@ -89,7 +78,9 @@ class DisciplinaController extends Controller
                             ->select('foto_disc')
                             ->get();
             foreach ($foto_antiguo as $foto_disc) {
-                    Storage::disk('foto_disc')->delete($foto_disc->foto_disc);    
+                if ($foto_disc->foto_disc != 'usuario-sin-foto.png') {
+                    Storage::disk('foto_disc')->delete($foto_disc->foto_disc);
+                }    
             }  
             $foto_nueva = $request->file('foto_disc');
                     $nombre_foto = time().'-'.$foto_nueva->getClientOriginalExtension();
@@ -135,8 +126,8 @@ class DisciplinaController extends Controller
                             ->select('foto_disc')
                             ->get();
             foreach ($foto_antiguo as $foto_disc) {
-                if ($foto_disc->foto_disc!='usuario-sin-foto.png') {
-                    Storage::disk('foto_disc')->delete($foto_disc->foto_disc);     
+        if ($foto_disc->foto_disc != 'usuario-sin-foto.png') {
+            Storage::disk('foto_disc')->delete($foto_disc->foto_disc);     
                 }
                    
         }
